requirements.txt
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
rich==13.9.2
SQLAlchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
